+++ src/view/components/Drawer/DrawerItem/style.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  contents: {
+    marginTop: 10,
+    marginBottom: 10
+  }
+});
